# Faster rendering of the debugger tree