… galata/test/documentation/debugger.test.ts-snapshots/debugger-run-documentation-linux.png
… galata/test/documentation/debugger.test.ts-snapshots/debugger-variables-documentation-linux.png
… galata/test/jupyterlab/debugger.test.ts-snapshots/image-debug-session-global-variables-jupyterlab-linux.png
… galata/test/jupyterlab/debugger.test.ts-snapshots/start-debug-session-script-variables-jupyterlab-linux.png
… galata/test/jupyterlab/debugger.test.ts-snapshots/start-debug-session-variables-jupyterlab-linux.png
… galata/test/jupyterlab/notebook-markdown.test.ts-snapshots/do-not-highlight-standalone-dollar-jupyterlab-linux.png