## References

Fixes #13694, as a side-effect fixes a bug with multiple loop buttons showing when hovering over nested branches.

| Firefox 108 | IQM [Q1 - Q3] | mean | min |
|---|---|---|---|
| before | 4261.1 [3760.6 – 4932.3] ms | 4326.3 ms | 2197 ms |
| after | 906.7 [838.4 – 1005.5] ms | 1240.8 ms | 780 ms |

| Chrome 108 | IQM [Q1 - Q3] | mean | min |
|---|---|---|---|
| before | 4306.2 [4181.5 – 4500] ms | 4609.7 ms | 4046.3 ms |
| after | 1482 [1406.8 – 1555.8] ms | 1685.8 ms | 1306.2 ms |

<details>

UI profiler with default settings for new debugger scenario (https://github.com/jupyterlab/ui-profiler/pull/25):

![Screenshot from 2023-01-04 21-20-17](https://user-images.githubusercontent.com/5832902/210652068-9506e036-fbfc-4c0f-b88b-c17059eeefe8.png)

</details>

Before

![Screenshot from 2023-01-04 21-33-44](https://user-images.githubusercontent.com/5832902/210654264-f8c849ae-5c1e-453f-87a5-2c2f89e5859e.png)

After

![Screenshot from 2023-01-04 21-34-14](https://user-images.githubusercontent.com/5832902/210654245-d70983f8-c7f6-4c5e-a1e1-bc9d1ab75ebf.png)


## Code changes

- [x] only ever render one loop button per entire tree
- [x] do not render the collapser icon when the item cannot be expanded
- [x] rename `VariablesComponent` to `VariablesBranch` (it was too confusing to navigate `VariableComponent` and `VariablesComponent` when one is nested in the other, would likely lead to CSS errors with time if left as-is)

## User-facing changes

| Before | After |
|---|---|
| ![before](https://user-images.githubusercontent.com/5832902/210655783-3fee2b0a-cf6b-496d-84f5-de103feeb70b.gif) | ![after](https://user-images.githubusercontent.com/5832902/210655816-9582f596-8f2b-4c7a-97b3-b1811e19f59a.gif)|



## Backwards-incompatible changes

Only in CSS styles