@krassowski 2023-01-05
please update galata snapshots